requirements.txt
google-cloud-storage==2.14.0
httpx==0.26.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.11
openai==1.8.0
pandas==2.1.4
pydantic==2.5.2
numpy==1.26.3
rank-bm25==0.2.2
streamlit==1.30.0
streamlit-authenticator==0.2.3
tiktoken==0.5.2
tqdm==4.66.1
# spacy==3.7.2
weaviate-client==4.4b7